rename server/ to api/
